Fixed invalid escape sequences.
Note that only some are for regular expressions (only one of them being reported as having an invalid escape sequence).
The others are strings (well, bytes) used in a different way.

It is very hard to see whether the magic mimetype changes are okay.
For the first one, starting with 8BPS, for x-psd, I have found it in /usr/share/mime on Ubuntu 18.04,
but even then I am not entirely sure it is correct.

I could not find the others in there, but they do seem safer, as they are actually the same.
For example:

    b'<smil\>' == br'<smil\>'